requirements.txt
flask>=2.0.0
PyPDF2>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
numpy
faiss-cpu # o faiss-gpu si tienes GPU compatible
sentence-transformers # Para generar embeddings localmente
langchain>=0.1.0
langchain-google-genai


